Fixed compatibility errors with latest pksm-core version
On a mission to try and add g1&2 functionality, the most fundamental
step is fixing interfacing with pksm-core as they already implemented
all the save-editing work. We also have a reference implementation in
PKSM.

The changes here allow compilation of the project, and are almost
identical to @SNBeast's advances 4 years ago, so many thanks to them!

Changelog:
 - Added "dummy(?) const" _PKSM_PERSONAL_FOLDER
 - Fixed compatibility for PKSM-Core@e50bb2f
 - Fixed compatibility for PKSM-Core@91cdfb9 (calling .data(), use of
 ranges, and other stuff).
 - Added size of buffer to *pksm::PKX::getPKM(...), can't find which
 commit.
 - Fixed compatibility for PKSM-Core@78ac9c9 (adaptimg validItems
 from mapping to array of pairs).

Majority of it referenced  and checked PKSM app, *huge* props to the
people mainting that codebase and the core logic. Also thanks a lot
to the original devs and again to SNBeast for their work.

Next up, testing and revising SNBeast's observations in later commits.