Nitrofs personals relocation (copy of SNBeast's work)
Takes more of SNBeast's work for updating to a later version of
PKSM-core, accomodating the change of the "personals" folder in
nitrofs. Unsure of the exact purpose of this functionality/change.

Props to SNBeast for the commits in #67, they are of great help,
especially with me not knowing a whole lot about the dev tools.