Testing and integration of #67
Testing and integration of the rest of the code written in #67.
The bugs have been tested, but not the solutions as of now, so more
work needs to be done, either by save inspection on real hardware or
by cross-referencing from the PKSM frontend.

Changelog:
 - Tested and confirmed bugs on bag.cpp and trainer.cpp, yet to test proposed solutions by SNBeast
 - de-hardcoded fwrite nmemb, congruent with how PKSM saves pokemon.
 - added code for injecting pokemon in gens 1&2.
 - fixed lv 1 not existing and hatching @ lv 5 in gen 2 and 1 as it happens w/ gen 3
 - enabled custom item id managing for < gen 3
 - Japanese name issues (corroborated w/ PKSM codebase)

Note: some other changes for the first commit of #67 were already merged in between
the original PR and the making of this branch. I'm also not changing the save from
a shared pointer into a unique pointer until I get the time and skill to check that
it doesn't break.